Add menu switch component
